Quick hack to focus the search box when viewed
The menu is provided by a jquery plugin called `dlmenu`.
Upon initialization, it can take an `onLevelClick` callback which will
fire after a sub-menu has been opened.
(I learned that from here:
https://github.com/codrops/ResponsiveMultiLevelMenu/blob/master/js/jquery.dlmenu.js#L29)

The first problem is that the plugin is initialized in
`source/javascripts/_main.js`
and then somehow is minified into
`source/javascripts/scripts.min.js`,
but `rake generate` does not do this.

So I could create a callback that focuses the searchbox and assign it at
initialization time, but I'd have to manually minify the code (until I
can work out how this actually happens), but this feels a little wrong
considering the callback happens every time any submenu is opened.
Calling this for every submenu would not cause an error, so it should be
ok, but it would be better if I could call my specific callback only for
the search submenu.

This change is a hack to only try to focus the searchbox when the search
sub-menu is clicked. It waits half a second for the element to be shown,
which is also dirty, but its the smallest change to get the desired
effect.

... One day, I should do this better :/