Hack to make the site search work properly!
I'm so confused! The old code (with a hidden 'q' input) works sometimes,
but not others - and it seems to be mostly random!

With the hidden input, most times, it ends up going to:
  https://www.google.com.au/search?q=site:blog.stevenocchipinti.com&q=mySearchTerm
This effectively does a search for `site:blog.stevenocchipinti.com mySearchTerm`

But sometimes, it ends up ignoring the first `q` parameter and just
going a normal google search without the extra `site:` filter.

It seems like on my home internet connection, having both `q` inputs is
fine, but on my work network/vpn, the search URL ends up being:
  https://www.google.com.au/search?q=searchTerm&gws_rd=cr&ei=gwBSVff3NMP38QWl14DIBA
(no `site:` filter)

This change is pretty hacky, but it ensures there is only ever one `q`
parameter and manually prepends the `site:` filter.

Another option would be to use https://cse.google.com.au, but this seems
a little heavy and I'm not sure how hard it would be to style.

This whole search thing should really be cleaned up one day :/
It also has a <style> tag that should probably go too!