Store travel thumbnails in the repo
Google images don't keep the URLs the same anymore (and they are really
long now). This isn't a reliable method anymore, although the older ones
from blogger still seem to be good.

(Kind of wish Google would offer this as a service as part of
photos.google.com, the image resizing, etc. is pretty handy)