Move the rest of the travel thumbnails to the repo
This is mainly to make it uniform.
If I eventually upgrade the image size to a bigger one, I could use this
commit to find the original Google URLs that I could change the size
parameter for to download new versions, or get the originals from my
photo albums.

The filenames include the size to bust the browser cache if/when I
upgrade to larger images.